muxa cosa en prim

+++ Prim.h
@@ -3,9 +3,7 @@
 #include "lag.h"
 
 void PRIM (int p, ertzPisuPos hzm []);
-void ErantsiErt(ertzPisuPos hzm [],int k,int Auzokide [],int PisuMin [],int sluz);
+void ErantsiErt(ertzPisuPos hzm [],int k,int Auzokide [],float PisuMin [],int sluz);
 void main2(void);
-
-
 
 #endif
